Create a base class separate from the Moq details.
