+++ src/AutoMoq/AutoMoq.csproj
@@ -73,6 +73,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Unity\AutoMockingBuilderStrategy.cs" />
     <Compile Include="Unity\AutoMockingContainerExtension.cs" />
+    <Compile Include="Unity\Base.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="License.txt" />
